# api-docs.json
{"swagger":"2.0","info":{"description":"Donor Management API","version":"0.0.1-SNAPSHOT","title":"Spring Boot Rest API","contact":{"name":"Somnath Musib","url":"medium.com/codefountain","email":"codefountain@gmail.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"activity-controller","description":"Activity Controller"}],"paths":{"/activity":{"get":{"tags":["activity-controller"],"summary":"Get all activity","operationId":"getAllActivityUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"post":{"tags":["activity-controller"],"summary":"Add activity to database","operationId":"createActivityUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"activity","description":"activity","required":true,"schema":{"$ref":"#/definitions/Activity"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/activity/delete/{id}":{"delete":{"tags":["activity-controller"],"summary":"Delete activity by id","operationId":"deleteActivityUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/activity/title":{"get":{"tags":["activity-controller"],"summary":"Get activity by title field","operationId":"getByTitleUsingGET","produces":["*/*"],"parameters":[{"name":"title","in":"query","description":"title","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/activity/{id}":{"get":{"tags":["activity-controller"],"summary":"Get activity by id","operationId":"getActivityByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["activity-controller"],"summary":"Update activity field","operationId":"updateActivityUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"activityDetails","description":"activityDetails","required":true,"schema":{"$ref":"#/definitions/Activity"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseEntity"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Activity":{"type":"object","properties":{"description":{"type":"string"},"endDateTime":{"type":"string"},"history":{"type":"array","items":{"$ref":"#/definitions/History"}},"id":{"type":"integer","format":"int64"},"info":{"type":"string"},"startDateTime":{"type":"string"},"summary":{"type":"string"},"title":{"type":"string"},"upChanges":{"$ref":"#/definitions/Activity"}},"title":"Activity"},"Changes":{"type":"object","properties":{"fieldName":{"type":"string"},"newValue":{"type":"string"},"oldValue":{"type":"string"}},"title":"Changes"},"History":{"type":"object","properties":{"changes":{"type":"array","items":{"$ref":"#/definitions/Changes"}},"dateTime":{"type":"string"},"type":{"type":"string","enum":["UPDATE","COMPOSE"]}},"title":"History"},"ResponseEntity":{"type":"object","properties":{"body":{"type":"object"},"statusCode":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","425 TOO_EARLY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"statusCodeValue":{"type":"integer","format":"int32"}},"title":"ResponseEntity"}}}